merge: merge VM into Compiler branch
- fix compilation for multiple executables
- refactor project structure
