+++ .gitignore
@@ -1,5 +1,5 @@
 Glados/.stack-work/
 Glados/stack.yaml.lock
 Glados/*~
-glados
+glados-*
 .vscode
