--- Glados/Setup.hs
@@ -1,2 +1,0 @@
-import Distribution.Simple
-main = defaultMain
