+++ Glados/app/MainCompiler.hs
@@ -1,6 +1,4 @@
-module Main (
-    main
-) where
+module Main where
 
 -- IMPORTS
